refactor: improve code quality and fix critical layout bugs
- Fix critical grid layout bug causing images to overflow viewport
- Remove direct DOM manipulation in favor of React patterns
- Organize modal components into dedicated modals/ folder
- Add filename truncation with tooltip for long names
- Update documentation (README, DEVELOPER_GUIDE, CLAUDE.md)

Technical improvements:
* Replace inline grid calculation with responsive Tailwind classes
* Remove unused handleScrollToCustomize function and DOM refs
* Move BulkRenameModal, CropModal, ImageSettingsModal to modals/
* Add text-overflow ellipsis with Mantine Tooltip for filenames
* Update all import paths to reflect new modal locations

Documentation:
* Add CLAUDE.md with AI assistant guidelines
* Update DEVELOPER_GUIDE.md to v2.0.0 with React 19 changes
* Refresh README.md with latest features and project structure
