+++ src/components/modals/ImageSettingsModal.tsx
